Move service interface loading into its own class
This cleans up the main server code by moving the stuff responsible
for loading service interfaces into a ServiceInterfaceLoader class.

ServiceInterfaceLoader is a class instead of an object because we need
a ContextShift to derive a coursier.util.Sync instance for IO. We get
the ContextShift provided by IOApp through the implicit constructor
argument.

In the future, the ServiceInterfaceLoader constructor will likely
require a configuration object that will control things like the cache
location and what resolvers to use.
